refactor: merge WorldSetupScreen into UniverseSetupScreen
