--- engine/src/main/resources/org/terasology/engine/assets/ui/worldSetupScreen.ui
@@ -1,119 +1,0 @@
-{
-  "type": "engine:worldSetupScreen",
-  "skin": "engine:mainMenu",
-  "contents": {
-    "type": "relativeLayout",
-    "contents": [
-      {
-        "type": "UILabel",
-        "id": "subtitle",
-        "family": "title",
-        "text": "${engine:menu#world-setup}",
-        "layoutInfo": {
-          "height": 48,
-          "position-horizontal-center": {},
-          "position-top": {
-            "target": "TOP",
-            "offset": 48
-          }
-        }
-      },
-      {
-        "type": "UIBox",
-        "id": "container",
-        "layoutInfo": {
-          "width": 520,
-          "position-horizontal-center": {},
-          "position-top": {
-            "target": "BOTTOM",
-            "offset": 16,
-            "widget": "subtitle"
-          },
-          "position-bottom": {
-            "target": "TOP",
-            "widget": "close",
-            "offset": 16
-          }
-        }
-      },
-      {
-        "type": "ColumnLayout",
-        "columns": 1,
-        "verticalSpacing": 16,
-        "horizontalSpacing": 8,
-        "layoutInfo": {
-          "width": 504,
-          "position-horizontal-center": {},
-          "position-top": {
-            "target": "TOP",
-            "widget": "container",
-            "offset": 8
-          },
-          "position-bottom": {
-            "target": "TOP",
-            "widget": "close",
-            "offset": 24
-          }
-        },
-        "contents": [
-          {
-            "type": "RelativeLayout",
-            "family": "description",
-            "contents": [
-              {
-                "type": "ScrollableArea",
-                "content": {
-                  "type": "propertyLayout",
-                  "id": "properties",
-                  "rowConstraints": "[min]"
-                },
-                "layoutInfo": {
-                  "position-horizontal-center": {},
-                  "position-top": {
-                    "target": "TOP"
-                  },
-                  "position-bottom": {
-                    "target": "TOP",
-                    "offset": 16,
-                    "widget": "apply"
-                  }
-                }
-              },
-              {
-                "type": "RowLayout",
-                "horizontalSpacing": 4,
-                "id": "apply",
-                "layoutInfo": {
-                  "position-horizontal-center": {},
-                  "height": 32,
-                  "width": 260,
-                  "position-top": {
-                    "target": "BOTTOM",
-                    "widget": "properties",
-                    "offset": 16
-                  }
-                },
-                "contents": [
-                ]
-              }
-            ]
-          }
-        ]
-      },
-      {
-        "type": "UIButton",
-        "text": "Done",
-        "id": "close",
-        "layoutInfo": {
-          "width": 128,
-          "height": 32,
-          "position-horizontal-center": {},
-          "position-bottom": {
-            "target": "BOTTOM",
-            "offset": 48
-          }
-        }
-      }
-    ]
-  }
-}
